add some tests and new functions

+++ src/main/java/br/com/emprestimo/domain/LoanPaymentsEntity.java
@@ -27,7 +27,7 @@ public class LoanPaymentsEntity {
     private UUID paymentId;
 
     @Column(name = "loan_value")
-    private BigDecimal paymentValue;
+    private Double paymentValue;
 
     @Column(name = "loan_payment_supposed_day")
     private LocalDate paymentSupposedPayDay;
